Trading Fix

+++ public/js/cardManager.js
@@ -77,9 +77,13 @@ function CardManager(playerObj, gameMap, ingameMenu, informationMenu, xStart, yS
 		
 		if(idx!=undefined)
 		{
-			ownedFields[idx][1][0].destroyChildren();
-			ownedFields[idx][1][0].clearCache();
-			ownedFields[idx][1][1].clearCache();
+			if(ownedFields[idx][1]!=undefined)
+			{
+				ownedFields[idx][1][0].destroyChildren();
+				ownedFields[idx][1][0].clearCache();
+				ownedFields[idx][1][1].clearCache();
+			}
+
 			ownedFields.splice(idx, 1);
 			reorder();
 			return true;
